refactor[#523] : entities/profile/api 타입 수정
- 외부로 export 되는 타입 type 폴더에서 정의
